+++ src/undoredo.h
@@ -47,8 +47,8 @@ public:
 	void clear() { clear(undo_); clear(redo_); }
 	void undo(T*& current) { doUndoRedo(undo_, redo_, current); }
 	void redo(T*& current) { doUndoRedo(redo_, undo_, current); }
-	uint undoSize() { return undo_.size(); }
-	uint redoSize() { return redo_.size(); }
+	size_t undoSize() { return undo_.size(); }
+	size_t redoSize() { return redo_.size(); }
 	void store (T* current) {
 		undo_.push_back(current->clone());
 		clear(redo_);
